% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{growthstandards-package}
\alias{growthstandards-package}
\alias{growthstandards}
\title{growthstandards: Anthropometric Growth Standard Calculations}
\description{
Convert anthropometric measurements to or from z-scores and centiles
using the WHO and INTERGROWTH growth standards. The WHO standard are used
for infants, children, and young adults. INTERGROWTH standards are
provided for fetal and newborn (including very preterm) calculations.
}
\details{
\url{http://hbgdki.github.io/growthstandards/}
}
\examples{
help(package = "growthstandards")
browseURL("http://hbgdki.github.io/growthstandards/")
}
