% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{ig_coefs}
\alias{ig_coefs}
\title{List of INTERGROWTH birth standard coefficients}
\source{
\url{https://intergrowth21.tghn.org}
}
\description{
A list of coefficients from the INTERGROWTH birth standard.
}
\examples{
head(ig_coefs$hcircm$Female)
}
\references{
International standards for newborn weight, length, and head circumference by gestational age and sex: the Newborn Cross-Sectional Study of the INTERGROWTH-21st Project
Villar, José et al.
The Lancet, Volume 384, Issue 9946, 857-868
}
\seealso{
\code{\link{igb_centile2value}}, \code{\link{igb_value2centile}}, \code{\link{igb_zscore2value}}, \code{\link{igb_value2zscore}}
}
\keyword{data}
