% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_igb.R
\name{igb_centile2value}
\alias{igb_centile2value}
\alias{igb_zscore2value}
\title{Convert birth measurements to INTERGROWTH z-scores/centiles (generic)}
\usage{
igb_centile2value(gagebrth, p = 50, var = "lencm", sex = "Female")

igb_zscore2value(gagebrth, z = 0, var = "lencm", sex = "Female")
}
\arguments{
\item{gagebrth}{gestational age at birth in days}

\item{p}{centile(s) to convert (must be between 0 and 100)}

\item{var}{the name of the measurement to convert ("lencm", "wtkg", "hcircm")}

\item{sex}{"Male" or "Female"}

\item{z}{z-score(s) to convert}
}
\description{
Convert birth measurements to INTERGROWTH z-scores/centiles (generic)
}
\note{
For gestational ages between 24 and 33 weeks, the INTERGROWTH very early preterm standard is used.
}
\examples{
# get 99th centile for Male birth weights across some gestational ages
igb_centile2value(232:300, 99, var = "wtkg", sex = "Male")
}
\references{
International standards for newborn weight, length, and head circumference by gestational age and sex: the Newborn Cross-Sectional Study of the INTERGROWTH-21st Project
Villar, José et al.
The Lancet, Volume 384, Issue 9946, 857-868

INTERGROWTH-21st very preterm size at birth reference charts. Lancet  2016 doi.org/10.1016/S0140-6736(16) 00384-6.
Villar, José et al.
}
