% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_igb.R
\name{igb_zscore2lencm}
\alias{igb_zscore2lencm}
\alias{igb_zscore2wtkg}
\alias{igb_zscore2hcircm}
\alias{igb_centile2lencm}
\alias{igb_centile2wtkg}
\alias{igb_centile2hcircm}
\title{Convert INTERGROWTH z-scores/centiles to birth measurements}
\usage{
igb_zscore2lencm(gagebrth, z = 0, sex = "Female")

igb_zscore2wtkg(gagebrth, z = 0, sex = "Female")

igb_zscore2hcircm(gagebrth, z = 0, sex = "Female")

igb_centile2lencm(gagebrth, p = 50, sex = "Female")

igb_centile2wtkg(gagebrth, p = 50, sex = "Female")

igb_centile2hcircm(gagebrth, p = 50, sex = "Female")
}
\arguments{
\item{gagebrth}{gestational age at birth in days}

\item{z}{z-score(s) to convert}

\item{sex}{"Male" or "Female"}

\item{p}{centile(s) to convert (must be between 0 and 100)}
}
\description{
Convert INTERGROWTH z-scores/centiles to birth measurements
}
\note{
For gestational ages between 24 and 33 weeks, the INTERGROWTH very early preterm standard is used.
}
\examples{
# get 99th centile for Male birth weights across some gestational ages
igb_centile2wtkg(168:300, 99, sex = "Male")

# recreate figure from preterm paper
d <- expand.grid(centile = c(3, 50, 97), gage = 168:300)
d$value <- igb_centile2lencm(d$gage, d$centile, sex = "Male")
lattice::xyplot(value ~ gage / 7, groups = centile, data = d, type = "l")
}
\references{
International standards for newborn weight, length, and head circumference by gestational age and sex: the Newborn Cross-Sectional Study of the INTERGROWTH-21st Project
Villar, José et al.
The Lancet, Volume 384, Issue 9946, 857-868

INTERGROWTH-21st very preterm size at birth reference charts. Lancet  2016 doi.org/10.1016/S0140-6736(16) 00384-6.
Villar, José et al.
}
