% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_igfet.R
\name{igfet_hccm2zscore}
\alias{igfet_hccm2zscore}
\alias{igfet_bpdcm2zscore}
\alias{igfet_ofdcm2zscore}
\alias{igfet_accm2zscore}
\alias{igfet_flcm2zscore}
\alias{igfet_hccm2centile}
\alias{igfet_bpdcm2centile}
\alias{igfet_ofdcm2centile}
\alias{igfet_accm2centile}
\alias{igfet_flcm2centile}
\title{Convert fetal ultrasound measurements to INTERGROWTH z-scores/centiles}
\usage{
igfet_hccm2zscore(gagedays, hccm)

igfet_bpdcm2zscore(gagedays, bpdcm)

igfet_ofdcm2zscore(gagedays, ofdcm)

igfet_accm2zscore(gagedays, accm)

igfet_flcm2zscore(gagedays, flcm)

igfet_hccm2centile(gagedays, hccm)

igfet_bpdcm2centile(gagedays, bpdcm)

igfet_ofdcm2centile(gagedays, ofdcm)

igfet_accm2centile(gagedays, accm)

igfet_flcm2centile(gagedays, flcm)
}
\arguments{
\item{gagedays}{gestational age in days}

\item{hccm}{head circumference (cm) measurement(s) to convert}

\item{bpdcm}{biparietel diameter (cm) measurement(s) to convert}

\item{ofdcm}{occipito-frontal diameter (cm) measurement(s) to convert}

\item{accm}{abdominal circumference (cm) measurement(s) to convert}

\item{flcm}{femur length (cm) measurement(s) to convert}
}
\description{
Convert fetal ultrasound measurements to INTERGROWTH z-scores/centiles
}
\examples{
# get centile for child at 100 gestational days with 11 cm head circumference
igfet_hccm2centile(100, 11)
}
\references{
International standards for fetal growth based on serial ultrasound measurements: the Fetal Growth Longitudinal Study of the INTERGROWTH-21st Project
Papageorghiou, Aris T et al.
The Lancet, Volume 384, Issue 9946, 869-879
}
