% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_igfet.R
\name{igfet_zscore2hccm}
\alias{igfet_zscore2hccm}
\alias{igfet_zscore2bpdcm}
\alias{igfet_zscore2ofdcm}
\alias{igfet_zscore2accm}
\alias{igfet_zscore2flcm}
\alias{igfet_centile2hccm}
\alias{igfet_centile2bpdcm}
\alias{igfet_centile2ofdcm}
\alias{igfet_centile2accm}
\alias{igfet_centile2flcm}
\title{Convert INTERGROWTH z-scores/centiles to fetal ultrasound measurements}
\usage{
igfet_zscore2hccm(gagedays, z = 0)

igfet_zscore2bpdcm(gagedays, z = 0)

igfet_zscore2ofdcm(gagedays, z = 0)

igfet_zscore2accm(gagedays, z = 0)

igfet_zscore2flcm(gagedays, z = 0)

igfet_centile2hccm(gagedays, p = 50)

igfet_centile2bpdcm(gagedays, p = 50)

igfet_centile2ofdcm(gagedays, p = 50)

igfet_centile2accm(gagedays, p = 50)

igfet_centile2flcm(gagedays, p = 50)
}
\arguments{
\item{gagedays}{gestational age in days}

\item{z}{z-score(s) to convert}

\item{p}{centile(s) to convert (must be between 0 and 100)}
}
\description{
Convert INTERGROWTH z-scores/centiles to fetal ultrasound measurements
}
\examples{
# get value for median head circumference for child at 100 gestational days
igfet_centile2hccm(100, 50)
}
\references{
International standards for fetal growth based on serial ultrasound measurements: the Fetal Growth Longitudinal Study of the INTERGROWTH-21st Project
Papageorghiou, Aris T et al.
The Lancet, Volume 384, Issue 9946, 869-879
}
