% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_gwg.R
\name{iggwg_value2zscore}
\alias{iggwg_value2zscore}
\alias{iggwg_value2centile}
\title{Convert mother's gestational weight gain (GWG) measurements to INTERGROWTH z-scores/centiles}
\usage{
iggwg_value2zscore(gagedays, val, bmi = NULL)

iggwg_value2centile(gagedays, val, bmi = NULL)
}
\arguments{
\item{gagedays}{gestational age(s) in days at which gestational weight gain was measured (must be same length as \code{val})}

\item{val}{value(s) of the gestational weight gain (kg gained since pregnancy) measurement(s) to convert}

\item{bmi}{optional value(s) of mother's pre-pregnancy BMI (kg / m^2) to check validity of GWG standard}
}
\description{
Convert mother's gestational weight gain (GWG) measurements to INTERGROWTH z-scores/centiles
}
\examples{
# get z-score and centile for mother at 17 weeks with 2 kg gestational weight gain
iggwg_value2zscore(17 * 7, 2)
iggwg_value2centile(17 * 7, 2)
}
\references{
Gestational weight gain standards based on women enrolled in the Fetal Growth Longitudinal Study of the INTERGROWTH-21st Project: a prospective longitudinal cohort study
Ismail, Leila Cheikh et al.
BMJ 2016;352:i555
}
