% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_gwg.R
\name{iggwg_zscore2value}
\alias{iggwg_zscore2value}
\alias{iggwg_centile2value}
\title{Convert INTERGROWTH z-scores/centiles to gestational weight gains (GWG)}
\usage{
iggwg_zscore2value(gagedays, z = 0, bmi = NULL)

iggwg_centile2value(gagedays, p = 50, bmi = NULL)
}
\arguments{
\item{gagedays}{gestational age(s) in days}

\item{z}{z-score(s) to convert}

\item{bmi}{optional value(s) of mother's pre-pregnancy BMI (kg / m^2) to check validity of GWG standard}

\item{p}{centile(s) to convert (must be between 0 and 100)}
}
\description{
Convert INTERGROWTH z-scores/centiles to gestational weight gains (GWG)
}
\examples{
# get value for median GWG for mother at 40 gestational weeks
iggwg_centile2value(40 * 7, 50)
}
\references{
Gestational weight gain standards based on women enrolled in the Fetal Growth Longitudinal Study of the INTERGROWTH-21st Project: a prospective longitudinal cohort study
Ismail, Leila Cheikh et al.
BMJ 2016;352:i555
}
