% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_igprepost.R
\name{igprepost_centile2value}
\alias{igprepost_centile2value}
\title{Convert postnatal growth measurements for preterm infants to INTERGROWTH z-scores/centiles (generic)}
\usage{
igprepost_centile2value(pmagedays, p = 50, var = "lencm", sex = "Female")
}
\arguments{
\item{pmagedays}{postmenstrual age in days}

\item{p}{centile(s) to convert (must be between 0 and 100)}

\item{var}{the name of the measurement to convert ("lencm", "wtkg", "hcircm")}

\item{sex}{"Male" or "Female"}
}
\description{
Convert postnatal growth measurements for preterm infants to INTERGROWTH z-scores/centiles (generic)
}
\examples{
# get median wtkg of Females from 27 to 64 weeks postmenstrual age
igprepost_centile2value(27:64 * 7, 50, var = "wtkg", sex = "Female")
}
\references{
Postnatal growth standards for preterm infants: the Preterm Postnatal Follow-up Study of the INTERGROWTH-21st Project.
Villar J, Giuliani F, Bhutta ZA, et al.
Lancet Glob Health 2015; 3: e681–91.
}
