% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_igprepost.R
\name{igprepost_value2centile}
\alias{igprepost_value2centile}
\alias{igprepost_value2zscore}
\title{Convert postnatal growth measurements for preterm infants to INTERGROWTH z-scores/centiles (generic)}
\usage{
igprepost_value2centile(pmagedays, val, var = "lencm", sex = "Female")

igprepost_value2zscore(pmagedays, val, var = "lencm", sex = "Female")
}
\arguments{
\item{pmagedays}{postmenstrual age in days}

\item{val}{the value(s) of the anthro measurement to convert}

\item{var}{the name of the measurement to convert ("lencm", "wtkg", "hcircm")}

\item{sex}{"Male" or "Female"}
}
\description{
Convert postnatal growth measurements for preterm infants to INTERGROWTH z-scores/centiles (generic)
}
\examples{
# get centile of 0.99kg 27 week postmenstrual age Male and 0.91kg 27 week postmenstrual age Female
igprepost_value2centile(27 * 7, c(0.99, 0.91), var = "wtkg",
  sex = c("Male", "Female"))
}
\references{
Postnatal growth standards for preterm infants: the Preterm Postnatal Follow-up Study of the INTERGROWTH-21st Project.
Villar J, Giuliani F, Bhutta ZA, et al.
Lancet Glob Health 2015; 3: e681–91.
}
