% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_igprepost.R
\name{igprepost_lencm2zscore}
\alias{igprepost_lencm2zscore}
\alias{igprepost_wtkg2zscore}
\alias{igprepost_hcircm2zscore}
\alias{igprepost_lencm2centile}
\alias{igprepost_wtkg2centile}
\alias{igprepost_hcircm2centile}
\title{Convert postnatal growth measurements for preterm infants to INTERGROWTH z-scores/centiles}
\usage{
igprepost_lencm2zscore(pmagedays, lencm, sex = "Female")

igprepost_wtkg2zscore(pmagedays, wtkg, sex = "Female")

igprepost_hcircm2zscore(pmagedays, hcircm, sex = "Female")

igprepost_lencm2centile(pmagedays, lencm, sex = "Female")

igprepost_wtkg2centile(pmagedays, wtkg, sex = "Female")

igprepost_hcircm2centile(pmagedays, hcircm, sex = "Female")
}
\arguments{
\item{pmagedays}{postmenstrual age in days}

\item{lencm}{length(cm) measurement(s) to convert}

\item{sex}{"Male" or "Female"}

\item{wtkg}{weight (kg) measurement(s) to convert}

\item{hcircm}{head circumference (cm) measurement(s) to convert}

\item{wlr}{weight-length ratio values(s) to convert}
}
\description{
Convert postnatal growth measurements for preterm infants to INTERGROWTH z-scores/centiles
}
\examples{
igprepost_lencm2zscore(64 * 7, 64.68, sex = "Female")
}
\references{
Postnatal growth standards for preterm infants: the Preterm Postnatal Follow-up Study of the INTERGROWTH-21st Project.
Villar J, Giuliani F, Bhutta ZA, et al.
Lancet Glob Health 2015; 3: e681–91.
}
