% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_igprepost.R
\name{igprepost_zscore2value}
\alias{igprepost_zscore2value}
\title{Convert z-scores to postnatal growth measurements for preterm infants (generic)}
\usage{
igprepost_zscore2value(pmagedays, z = 0, var = "lencm", sex = "Female")
}
\description{
Convert z-scores to postnatal growth measurements for preterm infants (generic)
}
