% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{panel.who}
\alias{panel.who}
\alias{panel.igb}
\alias{panel.igfet}
\alias{geom_growthstandard}
\alias{geom_who}
\alias{geom_igb}
\alias{geom_igfet}
\title{Utility functions for adding growth standard bands to ggplot2/lattice plots}
\usage{
panel.who(
  x,
  x_var = "agedays",
  y_var = "htcm",
  sex = "Female",
  p = c(1, 5, 25, 50),
  color = NULL,
  alpha = 0.15,
  center = FALSE,
  labels = TRUE,
  x_trans = identity,
  y_trans = identity
)

panel.igb(
  gagebrth,
  var = "lencm",
  sex = "Female",
  p = c(1, 5, 25, 50),
  color = NULL,
  alpha = 0.15,
  center = FALSE,
  labels = TRUE,
  x_trans = identity,
  y_trans = identity
)

panel.igfet(
  gagedays,
  var = "hccm",
  p = c(1, 5, 25, 50),
  color = "green",
  alpha = 0.15,
  center = FALSE,
  labels = TRUE,
  x_trans = identity,
  y_trans = identity
)

geom_growthstandard(
  mapping = NULL,
  data = NULL,
  x_seq,
  x_var = "agedays",
  y_var,
  sex = "Female",
  p = c(1, 5, 25, 50),
  shade = NULL,
  alpha = 0.15,
  center = FALSE,
  x_trans = identity,
  y_trans = identity,
  standard = "who",
  inherit.aes = TRUE
)

geom_who(...)

geom_igb(..., var = "lencm")

geom_igfet(..., var = "hccm", color = "green")
}
\arguments{
\item{x, x_seq}{value or vector of values that correspond to a measure defined by \code{x_var}. \code{x_seq} is used with geom_*}

\item{x_var}{x variable name (typically "agedays")}

\item{y_var}{y variable name (typically "htcm" or "wtkg")}

\item{sex}{"Male" or "Female"}

\item{p}{centiles at which to draw the growth standard band polygons (only need to specify on one side of the median)}

\item{color, shade}{optional color to use for bands (will use \code{sex} to determine if not specified). \code{shade} is used with geom_*}

\item{alpha}{transparency of the bands}

\item{center}{should the bands be centered around the median?}

\item{labels}{should the centiles be labeled? (not implemented)}

\item{x_trans}{transformation function to be applied to x-axis}

\item{y_trans}{transformation function to be applied to y-axis}

\item{gagebrth}{gestational age at birth in days (for igb plots)}

\item{var}{variable name for y axis for igb or igfet plots ("lencm", "wtkg", or "hcircm" for igb; "accm", "bpdcm", "flcm", "hccm", or "ofdcm" for igfet)}

\item{gagedays}{gestational age in days (for igfet plots)}

\item{data, mapping, inherit.aes}{supplied direclty to \code{ggplot2::layer}}

\item{standard}{standard name to use.  Either \code{"who"}, \code{"igb"}, or \code{"igfet"}}

\item{...}{items supplied direclty to \code{geom_growthstandard}}
}
\description{
Utility functions for adding growth standard bands to ggplot2/lattice plots
}
\examples{
\dontrun{
#### ggplot2

library(ggplot2)
p <- ggplot(data = subset(cpp, subjid == 8), aes(x = agedays, y = htcm)) +
  geom_who(x_seq = seq(0, 2600, by = 10), y_var = "htcm") +
  geom_point()

#### lattice

library(lattice)
xyplot(wtkg ~ agedays, data = subset(cpp, subjid == 8),
  panel = function(x, y, ...) {
    panel.who(x = seq(0, 2558, by = 30),
      sex = "Male", y_var = "wtkg", p = 100 * pnorm(-3:0))
    panel.xyplot(x, y, ...)
  },
  col = "black"
)

# look at Male birth lengths superposed on INTERGROWTH birth standard
# first we need just 1 record per subject with subject-level data
cppsubj <- subset(cpp, !duplicated(cpp$subjid))
xyplot(birthlen ~ jitter(gagebrth), data = subset(cppsubj, sex == "Male"),
  panel = function(x, y, ...) {
    panel.igb(gagebrth = 250:310, var = "lencm", sex = "Male")
    panel.points(x, y, ...)
  },
  col = "black", alpha = 0.75,
  xlab = "Gestational Age at Birth (days)", ylab = "Birth Length (cm)"
)
}
}
