% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{panel.zband}
\alias{panel.zband}
\alias{geom_zband}
\title{Utility functions for adding growth standard bands to ggplot2/lattice plots}
\usage{
panel.zband(x, z = -3:0, color = "green", alpha = 0.25)

geom_zband(obj, x, z = -3:0, color = "green", alpha = 0.25)
}
\arguments{
\item{x}{range on x axis that should be covered by bands}

\item{z}{z-scores at which to draw bands (only need to specify on one side of zero)}

\item{color}{color to use for bands}

\item{alpha}{transparency of the bands}

\item{obj}{ggplot2 object to add z bands to}
}
\description{
Utility functions for adding growth standard bands to ggplot2/lattice plots
}
\examples{

\dontrun{
library(ggplot2)
p <- ggplot(data = cpp, aes(x = jitter(agedays), y = haz))
geom_zband(p, x = seq(0, 2600, by = 10)) +
  geom_point()

library(lattice)
xyplot(haz ~ jitter(agedays), data = cpp,
  panel = function(x, y, ...) {
    panel.zband(x)
    panel.xyplot(x, y, ...)
  },
  col = "black", alpha = 0.5
)
}
}
