% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_subject_data}
\alias{get_subject_data}
\alias{get_time_data}
\title{Get subject-level or time-varying variables and rows of longitudinal data}
\usage{
get_subject_data(dat)

get_time_data(dat)
}
\arguments{
\item{dat}{data frame with longitudinal data}
}
\description{
Get subject-level or time-varying variables and rows of longitudinal data
}
