% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{cm2in}
\alias{cm2in}
\alias{in2cm}
\alias{lb2kg}
\alias{kg2lb}
\alias{days2years}
\alias{years2days}
\alias{days2months}
\alias{months2days}
\alias{months2years}
\alias{years2months}
\title{Unit conversion utility functions}
\usage{
cm2in(x)

in2cm(x)

lb2kg(x)

kg2lb(x)

days2years(x)

years2days(x)

days2months(x)

months2days(x)

months2years(x)

years2months(x)
}
\arguments{
\item{x}{value(s) to convert}
}
\description{
Unit conversion utility functions
}
