% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_who.R
\name{who_value2zscore}
\alias{who_value2zscore}
\alias{who_value2centile}
\title{Convert anthro measurements to WHO z-scores/centiles (generic)}
\usage{
who_value2zscore(
  x,
  y,
  x_var = "agedays",
  y_var = "lenhtcm",
  sex = "Female",
  data = NULL
)

who_value2centile(
  x,
  y,
  x_var = "agedays",
  y_var = "lenhtcm",
  sex = "Female",
  data = NULL
)
}
\arguments{
\item{x}{value or vector of values that correspond to a measure defined by \code{x_var}}

\item{y}{value or vector of values that correspond to a measure defined by \code{y_var}}

\item{x_var}{x variable name (typically "agedays") - see details}

\item{y_var}{y variable name (typically "htcm", "lencm", or "wtkg") - see details}

\item{sex}{"Male" or "Female"}

\item{data}{optional data frame that supplies any of the other variables provided to the function}
}
\description{
Compute z-scores or centiles with respect to the WHO growth standard for given values of x vs. y (typically x is "agedays" and y is a measure like "htcm").
}
\details{
for all supported pairings of \code{y_var} and \code{x_var} , type \code{names(who_coefs)}
}
\examples{
# z-scores
who_value2zscore(1670, in2cm(44))
who_value2zscore(1670, lb2kg(48), y_var = "wtkg")

who_value2centile(1670, in2cm(44))
who_value2centile(1670, lb2kg(48), y_var = "wtkg")

# add haz derived from WHO data and compare to that provided with data
cpp$haz <- who_value2zscore(x = agedays,  y = lencm, sex = sex, data = cpp)

# note that you can also do it this way
#' cpp$haz <- who_value2zscore(cpp$agedays, cpp$lencm, sex = cpp$sex)
}
\seealso{
\code{\link{who_centile2value}}, \code{\link{who_zscore2value}}
}
