% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_who.R
\name{who_wtkg2zscore}
\alias{who_wtkg2zscore}
\alias{who_htcm2zscore}
\alias{who_lenhtcm2zscore}
\alias{who_lencm2zscore}
\alias{who_bmi2zscore}
\alias{who_hcircm2zscore}
\alias{who_muaccm2zscore}
\alias{who_ssftmm2zscore}
\alias{who_tsftmm2zscore}
\alias{who_wtkg2centile}
\alias{who_htcm2centile}
\alias{who_lenhtcm2centile}
\alias{who_lencm2centile}
\alias{who_bmi2centile}
\alias{who_hcircm2centile}
\alias{who_muaccm2centile}
\alias{who_ssftmm2centile}
\alias{who_tsftmm2centile}
\title{Convert anthro measurements to WHO z-scores/centiles}
\usage{
who_wtkg2zscore(agedays, wtkg, sex = "Female")

who_htcm2zscore(agedays, htcm, sex = "Female")

who_lenhtcm2zscore(agedays, lenhtcm, sex = "Female")

who_lencm2zscore(agedays, lencm, sex = "Female")

who_bmi2zscore(agedays, bmi, sex = "Female")

who_hcircm2zscore(agedays, hcircm, sex = "Female")

who_muaccm2zscore(agedays, muaccm, sex = "Female")

who_ssftmm2zscore(agedays, ssftmm, sex = "Female")

who_tsftmm2zscore(agedays, tsftmm, sex = "Female")

who_wtkg2centile(agedays, wtkg, sex = "Female")

who_htcm2centile(agedays, htcm, sex = "Female")

who_lenhtcm2centile(agedays, lenhtcm, sex = "Female")

who_lencm2centile(agedays, lencm, sex = "Female")

who_bmi2centile(agedays, bmi, sex = "Female")

who_hcircm2centile(agedays, hcircm, sex = "Female")

who_muaccm2centile(agedays, muaccm, sex = "Female")

who_ssftmm2centile(agedays, ssftmm, sex = "Female")

who_tsftmm2centile(agedays, tsftmm, sex = "Female")
}
\arguments{
\item{agedays}{age in days}

\item{wtkg}{weight (kg) measurement(s) to convert}

\item{sex}{"Male" or "Female"}

\item{htcm}{height(cm) measurement(s) to convert}

\item{lenhtcm}{length or height(cm) measurement(s) to convert}

\item{lencm}{length(cm) measurement(s) to convert}

\item{bmi}{body-mass index measurement(s) to convert}

\item{hcircm}{head circumference (cm) measurement(s) to convert}

\item{muaccm}{mid-upper arm circumference (cm) measurement(s) to convert}

\item{ssftmm}{subscalpular skinfold (mm) measurement(s) to convert}

\item{tsftmm}{triceps skinfold (mm) measurement(s) to convert}
}
\description{
Convert anthro measurements to WHO z-scores/centiles
}
\examples{
cpp$haz <- who_value2zscore(x = agedays,  y = lencm, sex = sex, data = cpp)
}
